dongle: schematic: add fuse
Add a 100 mA fuse for USB power supply. This is somewhat optional as
the host side should also have a fuse, but nice since the host-side
fuse will have a higher current rating and we're not allowed to draw
more than 100 mA without previous negotiation according to the USB
specification.