dongle: add test point pads to J202
This connector servers a dual purpose:
- allow soldering on a connector for programming
- contact by micro jig for GND, Reset, VDD & SWDIO

To increase connection with the micro jig, there are now additional
pads.

Note: the pads for VDD and SWDIO are intentionally smaller than the pads
for reset and GND, because the neighbouring pads for VDD and SWDIO pads
are connected to SWDCLK and GND respectively, while the neighbouring
pads for reset and GND are not connected to a net.

Further changes:
- remove silk screen markings for J202, as they conflicted with the new
  pads (the footprint for J202 now differs from the footprint in the
  library; this is excluded from DRC)
- add pin 1 indicator circle in silk screen for J202